Separate parser in conditianl check file (functional)
